refactor: change layouts structure in order to not repeat code
